pyproject.toml
[tool.poetry]
name = "workwechat-interface"
version = "0.1.0"
description = ""
authors = ["xiebo <shieber@163.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.5"
requests = "^2.22.0"
PyYAML = "^5.1"

[tool.poetry.dev-dependencies]
coverage = "^4.5.4"
coveralls = "^1.8.2"
